requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.0
matplotlib==3.7.1
seaborn==0.12.2
slack-sdk==3.21.0
SQLAlchemy>=1.4.28,<2.0
apache-airflow==2.7.2
pyarrow==12.0.1
mysql-connector-python==8.0.33
